ai/requirements.txt
python-dotenv
mistralai
langdetect
